refactor(input): extract system utilities from input package into separate one
