+++ internal/input/input.go
@@ -1,50 +1,16 @@
-// Package input provides utility functions for user input and system checks.
-// It includes functions to check for root privileges, prompt the user for input
-// and validate prompted inputs.
+// Package input provides utility functions for user input. It includes
+// functions to prompt the user for input and validate prompted inputs.
 package input
 
 import (
 	"bufio"
 	"fmt"
 	"os"
-	"os/exec"
 	"slices"
 	"strings"
-	"syscall"
 
 	ierror "github.com/borisdvlpr/gotail/internal/error"
 )
-
-// CheckRoot checks if the current user has root privileges.
-// If not, it re-executes the script with sudo.
-// If the sudo command is not found or fails, an error is return.
-func CheckRoot() error {
-	if os.Geteuid() == 0 {
-		return nil
-	}
-
-	fmt.Println("Setup must be run as root. Re-executing with sudo...")
-
-	sudoPath, err := exec.LookPath("sudo")
-	if err != nil {
-		return fmt.Errorf("%w", err)
-	}
-
-	execPath, err := os.Executable()
-	if err != nil {
-		return fmt.Errorf("failed to get executable path: %w", err)
-	}
-
-	args := []string{"sudo", execPath}
-	args = append(args, os.Args[1:]...)
-
-	err = syscall.Exec(sudoPath, args, os.Environ())
-	if err != nil {
-		return fmt.Errorf("failed to execute sudo: %w", err)
-	}
-
-	return nil
-}
 
 // PromptUser prompts the user with the given prompt string and reads the input from stdin.
 // If allowedReplies is provided, it ensures the user's input is one of the allowed replies.
